Move eye_movement plugins into separate files
